Refactor db helper methods in db.js file to utils/dbHelper.js
