+++ controllers/auth.js
@@ -4,10 +4,10 @@ const {
   getSignedJwtToken,
   hashPassword,
 } = require('../utils/auth');
-const { db, parseSqlUpdateStmt } = require('../db/db');
+const { db } = require('../db/db');
 const asyncHandler = require('../middleware/async');
 const ErrorResponse = require('../utils/errorResponse');
-const { cleanseData } = require('../utils/dbHelper');
+const { cleanseData, parseSqlUpdateStmt } = require('../utils/dbHelper');
 const sendEmail = require('../utils/sendEmail');
 
 /**
