+++ controllers/listings.js
@@ -1,6 +1,6 @@
-const { db, parseSqlUpdateStmt } = require('../db/db');
+const { db } = require('../db/db');
 const asyncHandler = require('../middleware/async');
-const { cleanseData } = require('../utils/dbHelper');
+const { cleanseData, parseSqlUpdateStmt } = require('../utils/dbHelper');
 const ErrorResponse = require('../utils/errorResponse');
 const { hashDecode } = require('../utils/hashIdGenerator');
