+++ controllers/skills.js
@@ -1,6 +1,7 @@
-const { db, parseSqlUpdateStmt } = require('../db/db');
+const { db } = require('../db/db');
 const asyncHandler = require('../middleware/async');
 const ErrorResponse = require('../utils/errorResponse');
+const { parseSqlUpdateStmt } = require('../utils/dbHelper');
 
 /**
  * @desc    Get all skills
